Rename DTOLifeInsurance, DTOPropertyInsurance and DTOVehicleInsurance
