+++ backend/src/main/java/com/github/iskrendev/insuranceprogram/models/VehicleInsuranceDTO.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 @Builder
-public record DTOVehicleInsurance(String firstName,
+public record VehicleInsuranceDTO(String firstName,
                                   String familyName,
                                   String zipCode,
                                   String city,
